+++ CSJ.NET/CSJ.NET.csproj
@@ -47,6 +47,8 @@
     <Compile Include="CsjDeserializer.cs" />
     <Compile Include="CsjSerializer.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="StringExtensions.cs" />
+    <Compile Include="Utils.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="CSJ.NET.nuspec" />
